# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Quinn AI Personal Assistant API
  description: |
    Quinn is an AI-powered personal assistant that helps manage emails, banking, 
    food planning, life tasks, and more. All endpoints require authentication via 
    AWS Cognito JWT tokens.
  version: 1.0.0
  contact:
    name: Quinn Support
    email: support@quinn.ai
servers:
  - url: https://8aqsagpkp6.execute-api.eu-west-2.amazonaws.com/dev
    description: Development server
  - url: https://api.quinn.ai
    description: Production server (future)
security:
  - bearerAuth: []
...
paths:
...
  /food/grocery-lists:
    get:
      tags: [Food]
      summary: Get grocery lists
      responses:
        "200":
          description: Grocery lists retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroceryList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags: [Food]
      summary: Create grocery list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroceryListRequest"
      responses:
        "201":
          description: Grocery list created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroceryList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

  # Life
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: AWS Cognito JWT token
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            message: "Validation error"
            details: "Field 'title' is required"
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            message: "Unauthorized"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            message: "Internal server error"
  schemas:
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
        details:
          type: string
        code:
          type: string
...
    GroceryList:
      type: object
      properties:
        listId:
          type: string
        userId:
          type: string
        name:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              quantity:
                type: string
              checked:
                type: boolean
        createdAt:
          type: string
          format: date-time
    CreateGroceryListRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              quantity:
                type: string
...